Add files via upload
Steganography refers to the technique of hiding information within other non-secret data or mediums to avoid detection. Unlike encryption, which scrambles data to make it unreadable, steganography conceals the very existence of the message. This is often done in such a way that the carrier file (such as an image, audio, or video) appears unchanged to the naked eye or ear.